backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
django-redis==5.4.0
django-filter==23.5
Pillow==10.2.0
openai==1.12.0
python-magic==0.4.27
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
boto3==1.34.34
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0